backend/pyproject.toml
[project]
name = "me4brain"
version = "0.1.0"
description = "Me4BrAIn - Piattaforma Universale di Memoria Agentica API-First"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Fulvio Ventura" }]

dependencies = [
    # API Framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "sse-starlette>=2",
    # LangGraph Orchestration
    "langgraph>=0.2",
    "langgraph-checkpoint-postgres>=2",
    "langchain-core>=0.3",
    # Vector Database
    "qdrant-client>=1.12",
    # Graph Database
    "neo4j>=5.26",
    # Cache & Streams
    "redis>=5",
    # Embeddings
    "sentence-transformers>=3",
    "torch>=2.5",
    # Auth
    "pyjwt[crypto]>=2.9",
    "python-keycloak>=4",
    # Observability
    "structlog>=24",
    "prometheus-client>=0.21",
    "psutil>=6.1",
    # Utils
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
    "networkx>=3.4",
    "tenacity>=9.1.2",
    "pypdf>=6.6.2",
    "pdf2image>=1.17.0",
    "google-api-python-client>=2.188.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "duckduckgo-search>=8.1.1",
    "llama-index-core>=0.14.13",
    "llama-index-graph-stores-neo4j>=0.5.1",
    "llama-index-vector-stores-qdrant>=0.9.1",
    "llama-index-embeddings-huggingface>=0.6.1",
    # Sports APIs (free)
    "nba-api>=1.5",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "slowapi>=0.1.9",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "python-multipart>=0.0.9",
    "pymupdf>=1.26.7",
    "openpyxl>=3.1.5",
    # File system watching (per skill watcher)
    "watchdog>=4",
    # Job queue e scheduler
    "arq>=0.26",
    "croniter>=2",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "pandas-ta>=0.4.71b0",
    "nest-asyncio>=1.6.0",
    "yfinance>=1.1.0",
    "yahooquery>=2.3.7",
    "psycopg-binary>=3.3.3",
    "mlx-lm>=0.31.0",
    "curl-cffi>=0.13.0",
    "fastmcp>=0.4.0",
    "json-repair>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6",
    "hypothesis>=6.0",  # Property-based testing
    "ruff>=0.9",
    "mypy>=1.14",
    "pre-commit>=4",
    "httpx>=0.28",  # per testing FastAPI
]

llm = [
    # LLM Providers (opzionale, per chi vuole usarli)
    "openai>=1.60",
    "anthropic>=0.43",
    "google-generativeai>=0.8",
]

[project.scripts]
me4brain = "me4brain.api.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/me4brain"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = ["E501"]  # line too long (gestito da formatter)

[tool.ruff.lint.isort]
known-first-party = ["me4brain"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "neo4j.*",
    "qdrant_client.*",
    "sentence_transformers.*",
    "redis.*",
    "keycloak.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short --cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src/me4brain"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@(abc\\.)?abstractmethod",
]

[dependency-groups]
dev = [
    "hypothesis>=6.151.9",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "respx>=0.22.0",
]
